Rewrite UI in classic Views for old standalone Android watches
The Compose/Wear OS UI required API 26+ and the Wear support library,
which blocked installation on cheap standalone Android watches (Android
4.4/5.1, no Wear OS). Replaces the entire UI layer with plain Android
Views and drops every AndroidX/Compose dependency, lowering minSdk to 19.

- GameView: software-drawn watch face + NES frame + touch handling
  (ported geometry/hit-test/analog logic), double-buffered frame upload.
- MainActivity: classic Activity with a Views ROM picker (tap to open,
  long-press to delete) and the game screen; framework intents/permissions
  instead of androidx.activity.
- Audio: switch to the legacy AudioTrack constructor + 16-bit PCM (works
  back to Android 4.4); native renderAudio now emits int16. The modern
  Builder/float path was API 21-23 and would crash on a 5.1 watch.
- Manifest: uses-feature watch and the com.google.android.wearable
  uses-library are now required="false" so non-Wear devices can install.
- Pixel font moved to assets and loaded via Typeface.createFromAsset
  (no androidx.core dependency).
- Removed Compose build features and all dependencies; bump to v2.0.

Native core (cpu/ppu/apu/mapper) unchanged. Host-compiled the C++ again
to validate; on-device build via CI / Android Studio still required.

+++ android/app/src/main/java/com/watchemu/app/nes/NativeBridge.kt
@@ -14,8 +14,8 @@ object NativeBridge {
     external fun setButton(button: Int, pressed: Boolean)
     external fun isRomLoaded(): Boolean
 
-    /** Generate [count] audio samples into [out]; returns video frames completed. */
-    external fun renderAudio(out: FloatArray, count: Int): Int
+    /** Generate [count] 16-bit PCM samples into [out]; returns video frames completed. */
+    external fun renderAudio(out: ShortArray, count: Int): Int
 
     /** Copy the most recently completed frame (ARGB_8888) into [dst]. */
     external fun getFrameBuffer(dst: IntArray)
